requirements.txt
asgiref==3.5.0
bitly-api==0.3
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-classy-tags==3.0.1
django-cms==3.9.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-formtools==2.3
django-sekizai==3.0.1
django-treebeard==4.5.1
djangocms-admin-style==3.0.0
djangorestframework==3.13.1
idna==3.3
pyshorteners==1.0.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
